abstract out modal logic so each drop-down menu item only requires on modal file (it will conditionally render based on which setting prop is passed in)
